requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu
python-dotenv
tiktoken
